Extract spellchecker launch init into spellchecker.ts
